[DELETE][CHANGES]
Elimino app y se cambia el nombre de app_loop a app.py ;)
